Commit now manages connections within the server between databases.
The change to pgp_db assumes multiple connection strings within the db_connects file, with a structure defined as a js object with 'home', 'holding' and 'dev' connection strings.
The new `dbheader()` function in pgp_db allows us to pass in the header information to the function and then make the decision based on the presence of a `dest` element within the header. This then passes out the Database connection object.
Implementation in pgfunctions then calls in all the SQL files and stuff but doesn't call the connection until we're actually in the call itself, which lets us use the Request element, so we can get req.header().