Main js handling functions modified to accept POST/GET
This required some new management of the data objects as they get passed in. We use the `getparam()` function to pull in parameters from both POST and GET methods, and then merge them. The subsequent challenge was that in the neotoma API we explicitly call the parameters, here we have to map them from existing Tilia functions (lines 95-100).  Methods seem to work in general. Will do some more testing.