Needed to escape quotes if text was passed.
This is in response to a bug report by Brian Grivna. Strings were being passed with quotes, which we going into the SQL as literal quotes. Currently the script removes both paired single and double quotes that surround the entire string.

I'm not sure if this will result in any other kinds of unintended consequences, but should work for now.