Setting up a minor bug fix.
Not sure if this will fix the 500 errors we're seeing, but the hope is that by using a single conserved DB object throughout the server we're reducing the possibility of errors.
Local testing seems to indicate that things are working, but I'm running into issues with the fetchtest program. Will continue to work on this.