Adding query cache and more documentation
The query caching lets us use an internal object to store `pgp.QueryFiles` by their file paths. This way we don't create duplicate query file objects, which have been an issue since early on in this project.

The function still behaves the same, but we clear out the error, and we should have faster responses since we don't have to read in the file each time.